@@ -54,7 +54,7 @@ public class ClientOverlay extends JPanel implements MouseMotionListener, MouseL
         addLayer(new RuneMenu(this));
         addLayer(new QuickReplies(this));
         addLayer(new ChampionSuggestions(this));
-        addLayer(new TeamCompAnalysis(this));
+        addLayer(new EstimatedPositions(this));
 //        addLayer(new NotificationWindow(this));
         try {
             InputStream is = getClass().getResourceAsStream("/fonts/Beaufort-Bold.ttf");
