@@ -10,6 +10,8 @@ import org.slf4j.LoggerFactory;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.AccessDeniedException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 
 import static com.sun.jna.platform.win32.WinReg.HKEY_CURRENT_USER;
 
@@ -23,13 +25,18 @@ public class ShortcutUtils {
         try {
             String dir = PathUtils.getWorkingDirectory();
             File iconFile = new File(dir + File.separator + "icon.ico");
-            ShellLink sl = new ShellLink(fileName);
+            ShellLink sl = new ShellLink();
             ShellLinkHelper slh = new ShellLinkHelper(sl);
             sl.setIconLocation(iconFile.getAbsolutePath());
             sl.setWorkingDir(dir);
             sl.setName(linkName);
             String absolutePath = new File(directory, linkName + ".lnk").getAbsolutePath();
-            slh.saveTo(absolutePath);
+            Path targetPath = directory.toPath().toAbsolutePath();
+            String root = targetPath.getRoot().toString();
+            String path = targetPath.subpath(0, targetPath.getNameCount()).toString();
+            slh
+                    .setLocalTarget(root, path, ShellLinkHelper.Options.ForceTypeFile)
+                    .saveTo(absolutePath);
             log.info(String.format("Created shortcut for %s in %s", fileName, absolutePath));
         } catch (AccessDeniedException | ShellLinkException e) {
             log.error("An error occurred while creating a shortcut", e);
